This PR adds a struct that converts parquet rowgroups to arrow records. The conversion was previously stateless. With a struct, the door opens to a bunch of performance optimizations we can make, especially around avoiding redoing work for reach row group. In this PR, two optimizations are implemented:
- Reusing writers across row groups (only instantiated on row group schema change)
- Calculating index mappings upfront (on a row group schema change)

```
name          old time/op    new time/op    delta
QueryTypes-8    11.5ms ± 1%     5.3ms ± 1%  -53.78%  (p=0.008 n=5+5)
QueryMerge-8    19.5ms ± 0%    18.7ms ± 0%   -3.84%  (p=0.008 n=5+5)
QueryRange-8    12.5ms ±18%    11.3ms ± 2%   -9.42%  (p=0.016 n=5+4)

name          old alloc/op   new alloc/op   delta
QueryTypes-8    88.1MB ± 0%     5.6MB ± 0%  -93.63%  (p=0.008 n=5+5)
QueryMerge-8    56.9MB ± 0%    49.7MB ± 0%  -12.71%  (p=0.008 n=5+5)
QueryRange-8    26.1MB ± 0%    22.6MB ± 0%  -13.43%  (p=0.008 n=5+5)

name          old allocs/op  new allocs/op  delta
QueryTypes-8      136k ± 0%       93k ± 0%  -31.64%  (p=0.008 n=5+5)
QueryMerge-8      149k ± 0%      146k ± 0%   -1.95%  (p=0.008 n=5+5)
QueryRange-8     52.2k ± 0%     49.2k ± 0%   -5.64%  (p=0.016 n=5+4)
```

Fixes https://github.com/parca-dev/parca/issues/1689