@asubiotto 2022-09-14
Not sure why I'm hitting this in CI again: `race: limit on 8128 simultaneously alive goroutines is exceeded, dying`. I haven't upgraded parquet in this version or touched concurrency.