@asubiotto 2022-09-14
Ah, I see. Looks like parquet was bumped in 9a677b9c3c3a1b2359c9d221c712b10f002a8b5d with no failures. Might be a successful flake. A fix is in the works in any case (https://github.com/segmentio/parquet-go/pull/337).